data and model transformation functions added
