+++ src/metldata/builtin_transformations/count_content_values/data_transform.py
@@ -16,51 +16,14 @@
 """Data transformation logic for count content values transformation."""
 
 from collections import Counter
+from typing import Any
 
-from schemapack._internals.spec.custom_types import ResourceId
-from schemapack._internals.spec.datapack import ResourceIdSet
 from schemapack.spec.datapack import DataPack
 
-from metldata.builtin_transformations.add_content_properties.path import (
-    resolve_data_object_path,
-)
-from metldata.builtin_transformations.common.custom_types import (
-    MutableClassResources,
-    MutableDatapack,
-    MutableResource,
-    MutableResourceContent,
-    ResolveRelations,
-)
-from metldata.builtin_transformations.common.path.path import RelationPath
-from metldata.builtin_transformations.common.resolve_path import resolve_path
-from metldata.builtin_transformations.common.utils import data_to_dict
+from metldata.builtin_transformations.common.data_transform import transform_data
 from metldata.builtin_transformations.count_content_values.instruction import (
     CountContentValueInstruction,
 )
-from metldata.transform.exceptions import (
-    EvitableTransformationError,
-)
-
-
-def get_class_resources(
-    *, data: MutableDatapack, class_name: str
-) -> MutableClassResources:
-    """Extract the resources of a given class from the dictionary."""
-    resources = data["resources"].get(class_name)
-    if not resources:
-        raise EvitableTransformationError()
-    return resources
-
-
-def _resolve_relations(data: DataPack) -> ResolveRelations:
-    """Retains resolve_path function's access to a datapack argument."""
-
-    def partial_resolve_relations(
-        source_resource_id: ResourceId, path: RelationPath
-    ) -> frozenset[ResourceId]:
-        return resolve_path(data=data, source_resource_id=source_resource_id, path=path)
-
-    return partial_resolve_relations
 
 
 def count_content(
@@ -69,99 +32,13 @@ def count_content(
     instructions_by_class: dict[str, list[CountContentValueInstruction]],
 ) -> DataPack:
     """Apply all count content value transformation instructions."""
-    modified_data = data_to_dict(data)
-
-    for class_name, instructions in instructions_by_class.items():
-        transform_class(
-            class_name=class_name,
-            data=modified_data,
-            instructions=instructions,
-            resolve_relations=_resolve_relations(data=data),
-        )
-
-    return DataPack.model_validate(modified_data)
-
-
-def transform_class(
-    *,
-    class_name: str,
-    data: MutableDatapack,
-    instructions: list[CountContentValueInstruction],
-    resolve_relations: ResolveRelations,
-):
-    """Apply the count content value transformations to the specified class."""
-    # the target prefix refers to resources that will be modified by the transformation
-    target_resources = get_class_resources(data=data, class_name=class_name)
-
-    for instruction in instructions:
-        path = instruction.source.relation_path
-        referenced_class = path.target
-
-        # get resources for the class referenced by the relation path
-        referenced_resources = get_class_resources(
-            data=data, class_name=referenced_class
-        )
-        for resource_id, resource in target_resources.items():
-            relation_target_ids = resolve_relations(
-                resource_id,
-                path,
-            )
-            transform_resource(
-                referenced_resources=referenced_resources,
-                target_resource=resource,
-                relation_target_ids=relation_target_ids,
-                instruction=instruction,
-            )
-
-
-def transform_resource(
-    *,
-    referenced_resources: MutableClassResources,
-    target_resource: MutableResource,
-    relation_target_ids: frozenset[str],
-    instruction: CountContentValueInstruction,
-):
-    """Apply the count content value transformation to each resource of a class."""
-    target_content = target_resource.get("content")
-    if target_content is None:
-        raise EvitableTransformationError()
-
-    values_to_count = get_values_to_count(
-        relation_target_ids=relation_target_ids,
-        referenced_resources=referenced_resources,
-        content_path=instruction.source.content_path,
+    return transform_data(
+        data=data,
+        instructions_by_class=instructions_by_class,
+        calculate_value=occurrence_counter,
     )
-    target_object = get_modification_target(
-        data=target_content, instruction=instruction
-    )
-    target_property = instruction.target_content.property_name
-    target_object[target_property] = dict(Counter(values_to_count))
 
 
-def get_values_to_count(
-    *,
-    relation_target_ids: ResourceId | ResourceIdSet,
-    referenced_resources: MutableClassResources,
-    content_path: str,
-):
-    """Get countable properties from all resources referred to by the relation."""
-    try:
-        return [
-            referenced_resources[resource_id]["content"].get(content_path)
-            for resource_id in relation_target_ids
-        ]
-    except KeyError as exc:
-        raise EvitableTransformationError() from exc
-
-
-def get_modification_target(
-    *, data: MutableResourceContent, instruction: CountContentValueInstruction
-):
-    """Get the json object that is to be modified."""
-    path = instruction.target_content.object_path
-    property = instruction.target_content.property_name
-
-    target = resolve_data_object_path(data=data, path=path)
-    if not isinstance(target, dict) or property in target:
-        raise EvitableTransformationError()
-    return target
+def occurrence_counter(values_to_count: list[Any]) -> dict[Any, int]:
+    """Count the occurrences of objects in a list."""
+    return dict(Counter(values_to_count))
