+++ src/metldata/builtin_transformations/sum_operation/main.py
@@ -1,12 +1,12 @@
 # Copyright 2021 - 2024 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 # for the German Human Genome-Phenome Archive (GHGA)
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +14,6 @@
 # limitations under the License.
 
 """A transformation to sum up the values of content properties."""
-
 
 from schemapack.spec.datapack import DataPack
 from schemapack.spec.schemapack import SchemaPack
@@ -33,7 +32,9 @@ from metldata.transform.base import DataTransformer, TransformationDefinition
 
 
 class SumOperationDataTransformer(DataTransformer[SumOperationConfig]):
-    """TODO"""
+    """A transformer that sums the values of a specified property within an object and
+    adds them to the content of a target object.
+    """
 
     def transform(self, data: DataPack) -> DataPack:
         """Transforms data.
