+++ tests/fixtures/example_transformations/sum_operation/single/transformed.schemapack.yaml
@@ -27,11 +27,7 @@ classes:
                             "size":
                               {
                                 "oneOf":
-                                  [
-                                    { "type": "number" },
-                                    { "type": "integer" },
-                                    { "type": "boolean" },
-                                  ],
+                                  [{ "type": "number" }, { "type": "boolean" }],
                               },
                           },
                       },
